refactor testing for suggest easy meals use case and identify iraqi meals use case
