+++ src/test/kotlin/logic/useCases/GetIraqiMealsUseCaseTest.kt
@@ -3,6 +3,7 @@ package logic.useCases
 import com.google.common.truth.Truth.assertThat
 import com.thechance.logic.MealsRepository
 import com.thechance.logic.useCases.GetIraqiMealsUseCase
+import helper.createIraqiMeal
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.jupiter.api.BeforeEach
@@ -77,7 +78,7 @@ class GetIraqiMealsUseCaseTest {
         val result = getIraqiMealsUseCase.getIraqiMeals()
 
         //Then
-        assertThat(result).hasSize(0)
+        assertThat(result).isEmpty()
     }
 
     @Test
@@ -101,35 +102,27 @@ class GetIraqiMealsUseCaseTest {
     @Test
     fun `should return only iraqi meals when multiple (mixed) meals are given`() {
         // Given
+        val iraqiMeal1 = createIraqiMeal(name = "Pizza", tags = listOf("main-dish", "iraqi"), description = "A new iraqi meal")
+        val nonIraqiMeal1 = createIraqiMeal(name = "barmia", description = "yummy, inexpensive main dish!", tags = listOf("60-minutes-or-less", "meat"))
+        val nonIraqiMeal2 = createIraqiMeal(name = "coconut bread", description = "found this recipe on a french recipe website.", tags = listOf("low-protein", "cupcakes"))
+        val iraqiMeal2 = createIraqiMeal(name = "Pizza", tags = listOf("main-dish", "iraqi"), description = "A new meal")
+
         every { mealsRepository.getAllMeals() } returns listOf(
-            createIraqiMeal(
-                name = "Pizza",
-                tags = listOf("main-dish", "iraqi"),
-                description = "A new iraqi meal"
-            ),
-            createIraqiMeal(
-                name = "barmia",
-                description = "yummy, inexpensive main dish!",
-                tags = listOf("60-minutes-or-less", "meat")
-            ),
-            createIraqiMeal(
-                name = "coconut bread",
-                description = "found this recipe on a french recipe website.",
-                tags = listOf("low-protein", "cupcakes")
-            ),
-            createIraqiMeal(
-                name = "Pizza",
-                tags = listOf("main-dish", "iraqi"),
-                description = "A new meal"
-            )
+            iraqiMeal1,
+            nonIraqiMeal1,
+            nonIraqiMeal2,
+            iraqiMeal2
         )
 
-        //When
+        // When
         val result = getIraqiMealsUseCase.getIraqiMeals()
 
-        //Then
-        assertThat(result).hasSize(2)
+        // Then
+        assertThat(result)
+            .containsExactly(iraqiMeal1, iraqiMeal2)
+
     }
+
 
     @Test
     fun `should return null when meal is tagged with (iraq) not (iraqi)`() {
@@ -146,7 +139,7 @@ class GetIraqiMealsUseCaseTest {
         val result = getIraqiMealsUseCase.getIraqiMeals()
 
         //Then
-        assertThat(result).hasSize(0)
+        assertThat(result).isEmpty()
     }
 
     @Test
@@ -158,7 +151,7 @@ class GetIraqiMealsUseCaseTest {
         val result = getIraqiMealsUseCase.getIraqiMeals()
 
         //Then
-        assertThat(result).hasSize(0)
+        assertThat(result).isEmpty()
     }
 
     @Test
